refactor: extract picker component
